# [Perf] Cache key size in CngKey