An internal partner is reporting that they're seeing CNG signature validation latency show up as a bottleneck. The culprit seems to be that each signing \& validation operation queries the `CngKey.KeySize` property.

https://github.com/dotnet/runtime/blob/5d7b16f3ccc2680dde0270b2190ab4adb0f813b7/src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/RSACng.Key.cs#L20-L26

https://github.com/dotnet/runtime/blob/5d7b16f3ccc2680dde0270b2190ab4adb0f813b7/src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/CngAlgorithmCore.cs#L47-L56

The `CngKey.KeySize` property is not cached anywhere in the system, which means that each query to this property results in an LRPC call back into lsass or lsaiso. This context switch is what appears to be introducing the latency.

Since `CngKey` instances can't change their key size once created, we can cache the key size on our end without requiring calls back through the `NCrypt*` (lsass) layer. That significantly reduces the latency.

```cs
public class CngRunner
{
    private RSACng _rsa;
    private ECDsaCng _ecdsa;

    private static byte[] _emptyDigest = new byte[256 / 8];
    private byte[] _rsaSignedHash;
    private byte[] _ecdsaSignedHash;

    [GlobalSetup]
    public void Setup()
    {
        _rsa = new RSACng(2048);
        _ecdsa = new ECDsaCng(256);

        _rsaSignedHash = _rsa.SignHash(_emptyDigest, HashAlgorithmName.SHA256, RSASignaturePadding.Pss);
        if (!Rsa_VerifyHash()) { throw new CryptographicException("Self-test failed."); }

        _ecdsaSignedHash = _ecdsa.SignHash(_emptyDigest);
        if (!Ecdsa_VerifyHash()) { throw new CryptographicException("Self-test failed."); }
    }

    [Benchmark]
    public bool Rsa_VerifyHash() => _rsa.VerifyHash(_emptyDigest, _rsaSignedHash, HashAlgorithmName.SHA256, RSASignaturePadding.Pss);

    [Benchmark]
    public bool Ecdsa_VerifyHash() => _ecdsa.VerifyHash(_emptyDigest, _ecdsaSignedHash);
}
```

|           Method |        Job |    Toolchain |      Mean |    Error |   StdDev | Ratio |
|----------------- |----------- |------------- |----------:|---------:|---------:|------:|
|   Rsa_VerifyHash | Job-EBVYOU | cngkey_cache |  48.03 μs | 0.140 μs | 0.131 μs |  0.38 |
|   Rsa_VerifyHash | Job-QBWHWJ |         main | 125.42 μs | 0.505 μs | 0.394 μs |  1.00 |
|                  |            |              |           |          |          |       |
| Ecdsa_VerifyHash | Job-EBVYOU | cngkey_cache | 220.13 μs | 2.552 μs | 2.731 μs |  0.75 |
| Ecdsa_VerifyHash | Job-QBWHWJ |         main | 295.14 μs | 0.885 μs | 0.739 μs |  1.00 |

### Open discussion

Technically this is a behavioral change, since it means that the `CngKey.KeySize` property will remain accessible on an otherwise disposed instance, just as long as it has already been fetched at least once prior to disposal. I could easily add a disposed check (and unit test to validate this behavior) if we think preserving the earlier behavior is worthwhile.