@GrabYourPitchforks 2023-07-28
Updated the PR to use -1 instead of 0 as a sentinel, plus moved some logic around to account for early return scenarios. I also added a unit test to ensure that all non-handle properties throw after the CngKey instance is disposed, even if the properties had been previously accessed and cached. This should help thwart behavioral regressions.

Note to reviewers: I suggest **hiding whitespace changes** when reviewing the latest iteration, since some indentation changed within the KeySize property getter.

Perf results from this build are nearly identical to the earlier run. I won't bother posting them.