@brianrob 2023-07-28
A bit of justification here: This isn't a big deal for one off usages, but any time there is a system that is caching an instance for use over and over, the LRPC call is a significant bottleneck, and limits the scale of usage to be what lsass can handle.  This is a common situation in cloud services, so a change to cache the key size is worthwile.