@vcsjones 2023-07-28
@brianrob out of curiosity, are you using RSACng directly? Or are you using RSA.Create?

If the former, are you using CNG specific features that would exclude using RSA.Create?