@brianrob 2023-07-28
@vcsjones, I'm one of the folks that was analyzing perf data, but don't work on the service itself.  I've sent mail asking them to reply here.