@vcsjones 2023-07-28
Gotcha. I only ask because starting in .NET 8, `RSA.Create()` will use bcrypt, not ncrypt, when possible. So all the RPC overhead goes away entirely.