@GrabYourPitchforks 2023-07-28
Internal partner reports that they're using `RSACertificateExtensions.GetRSAPublicKey` extension method. On net8, this should go through bcrypt instead of ncrypt, so you're correct that the perf overhead should disappear.

I've changed the benchmark to perform signing instead of verification, since private key operations still require bouncing through lsass. The perf gains are still measurable: around 100μs per op.

|         Method |        Job |    Toolchain |     Mean |   Error |  StdDev | Ratio |
|--------------- |----------- |------------- |---------:|--------:|--------:|------:|
|   Rsa_SignHash | Job-QLWHCM | cngkey_cache | 659.9 μs | 5.05 μs | 4.47 μs |  0.88 |
|   Rsa_SignHash | Job-OFFVFC |         main | 748.6 μs | 2.14 μs | 1.90 μs |  1.00 |
|                |            |              |          |         |         |       |
| Ecdsa_SignHash | Job-QLWHCM | cngkey_cache | 373.1 μs | 2.74 μs | 2.57 μs |  0.78 |
| Ecdsa_SignHash | Job-OFFVFC |         main | 479.1 μs | 3.02 μs | 2.52 μs |  1.00 |