@vcsjones 2023-07-28
> The perf gains are still measurable: around 100μs per op.

Sure. I’m not trying to argue against the change - we should take it regardless since it has already demonstrated some improvements elsewhere. Just noting that the internal partner is likely to get an even more significant performance improvement for .NET 8 once they start going through bcrypt.