@bartonjs 2023-07-28
Really this shows that Windows should add similar caching.  But even if they did it'd save on the machinery of a P/Invoke.