@vcsjones 2023-07-28
> Really this shows that Windows should add similar caching.

Well... the key size can change - before NCryptFinalize is called. We (mostly) don't have that problem here. But I'm not so sure how reasonable it would be for Windows to add caching with that in mind.